Split Device helpers into .hpp/.cpp pairs, extract initBle() and publishBootMessage()
Move implementations from header-only modules into .cpp files:
- BootHelpers, BootMessage (new), ConfigUpdate, MqttCommands,
  SyncPublisher, TelemetryTask

Extract from startDevice():
- initBle() encapsulates CONFIG_BT_NIMBLE_ENABLED platform conditional
- publishBootMessage() with InitState enum moves to new BootMessage module

Fix include-order issues exposed by the .hpp/.cpp split:
- Log.hpp must precede headers that transitively include NvsStore.hpp
  (LOGGING_TAG macro dependency)
- firmwareVersion definition placed immediately before HttpUpdate.hpp /
  CrashManager.hpp (the only headers that reference it inline)
- Add missing mqtt namespace to TelemetryTask.hpp
- Add missing UpdateFilter.hpp include to Device.hpp
- Add missing ConfigBootPlan.hpp include to ConfigUpdate.cpp
- Add missing LogJson.hpp include to DeviceConfiguration.hpp

Mark Strings.hpp functions inline to fix ODR violations now that the
devices component has multiple translation units.
