+++ components/devices/CMakeLists.txt
@@ -1,7 +1,10 @@
 idf_component_register(
+    SRC_DIRS
+        src
     INCLUDE_DIRS
         src
     REQUIRES
+        functions
         kernel
         peripherals
 )
