+++ components/devices/src/BootMessage.cpp
@@ -1,0 +1,83 @@
+#include <Log.hpp>
+#include <BootMessage.hpp>
+
+#include <esp_app_desc.h>
+
+#include <KernelStatus.hpp>
+
+// CrashManager.hpp references firmwareVersion inline, so it must be defined before the include
+static const char* const firmwareVersion = reinterpret_cast<const char*>(esp_app_get_description()->version);
+#include <CrashManager.hpp>
+
+using namespace std::chrono;
+using namespace cornucopia::ugly_duckling::kernel;
+
+/**
+ * BOOT carries diagnostics and per-peripheral/function error feedback, but no configuration
+ * bodies (docs/Configuration.md, "BOOT, SYNC, UPDATE") -- fingerprints are reported separately
+ * by SYNC (initSyncTask), gated on kernelReady. `rejection` is present only when a
+ * requested-set revert (this boot or an earlier, unreported one) left one recorded. Published at
+ * QoS 2, consistent with every other outbound topic (sync, log, responses) -- also closes off the
+ * same duplicate-resend risk described on initTelemetryPublishTask, even though BOOT's own
+ * payload has no delta/counter state that a duplicate would corrupt.
+ */
+void publishBootMessage(
+    const std::shared_ptr<MqttRoot>& mqttRoot,
+    esp_reset_reason_t resetReason,
+    const std::string& macAddress,
+    const std::shared_ptr<NetworkConfig>& networkConfig,
+    InitState initState,
+    const JsonArray& peripheralsInitJson,
+    const JsonArray& functionsInitJson,
+    const std::shared_ptr<PowerManager>& powerManager,
+    const std::shared_ptr<DeviceDefinition>& deviceDefinition,
+    const std::optional<HardwareVersion>& hardwareVersion,
+    const std::optional<RejectionCode>& rejectionToReport,
+    const std::optional<RejectionCode>& firmwareDownloadRejection,
+    const std::optional<RollbackDetection>& rollback) {
+    mqttRoot->publish(
+        "boot",
+        [resetReason, macAddress, networkConfig, initState, peripheralsInitJson, functionsInitJson, powerManager, deviceDefinition, hardwareVersion, rejectionToReport, firmwareDownloadRejection, rollback](JsonObject& json) {
+            json["model"] = deviceDefinition->model;
+            json["revision"] = deviceDefinition->revision;
+            json["platform"] = UD_PLATFORM;
+            json["instance"] = networkConfig->instance.get();
+            json["mac"] = macAddress;
+            if (hardwareVersion.has_value()) {
+                json["batch"] = hardwareVersion->batch;
+                json["serial"] = hardwareVersion->serial;
+            }
+            json["version"] = firmwareVersion;
+#ifdef UD_DEBUG
+            json["debug"] = true;
+#else
+            json["debug"] = false;
+#endif
+            json["reset"] = resetReason;
+            json["wakeup"] = esp_sleep_get_wakeup_causes();
+            json["bootCount"] = bootCount++;
+            json["time"] = duration_cast<seconds>(system_clock::now().time_since_epoch()).count();
+            json["state"] = static_cast<int>(initState);
+            json["peripherals"].to<JsonArray>().set(peripheralsInitJson);
+            json["functions"].to<JsonArray>().set(functionsInitJson);
+            json["sleepWhenIdle"] = powerManager->sleepWhenIdle;
+            if (rejectionToReport) {
+                json["rejection"] = static_cast<int>(*rejectionToReport);
+            }
+            // firmwareRejection: either a failed download or a rollback (at most one)
+            if (rollback) {
+                json["firmwareRejection"] = static_cast<int>(rollback->rejectionCode);
+            } else if (firmwareDownloadRejection) {
+                json["firmwareRejection"] = static_cast<int>(*firmwareDownloadRejection);
+            }
+
+            // When a rollback was detected, attribute any coredump to the failed partition's
+            // version rather than the currently running (rolled-back-to) firmwareVersion
+            std::optional<std::string> rolledBackFromVersion;
+            if (rollback) {
+                rolledBackFromVersion = rollback->failedVersion;
+            }
+            CrashManager::handleCrashReport(json, rolledBackFromVersion);
+        },
+        Retention::NoRetain, QoS::ExactlyOnce, 5s);
+}
